pydanticai_agent_mcp/pyproject.toml
[project]
name = "pydanticai-agent-mcp"
version = "0.1.0"
description = "pydanticai agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=0.3.6",
    "python-dotenv>=1.1.1",
]
